#64 Refactor mock response into separate class
